Move tests from the lib.rs to a dedicated tests module
